#145: Implement `studyroom` command.
