+++ .env.example
@@ -14,6 +14,7 @@ SERVER_ID =
 
 ## Category IDs
 CATEGORY_ID_GAMING_ROOMS =
+CATEGORY_ID_STUDY_ROOMS =
 
 ## Channel IDs
 CHANNEL_ID_BOT =
